+++ desktop/infra/under-construction/index.html
@@ -5,8 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Under Construction - Classic 90s style project page with modern Mac aesthetics">
     <title>Under Construction | Paul Junsuk Han</title>
-    <link rel="stylesheet" href="../../styles/main.css">
-    <link rel="stylesheet" href="../../styles/construction.css">
+    <link rel="stylesheet" href="/styles/global.css">
+    <link rel="stylesheet" href="/styles/construction.css">
 </head>
 <body>
     <!-- Add mobile menu toggle button -->
@@ -24,19 +24,19 @@
 
     <!-- Desktop Icons -->
     <div class="desktop-icons">
-        <a href="../../" class="desktop-icon">
+        <a href="/" class="desktop-icon">
             <div class="icon-image">💾</div>
             <div class="icon-label">Home</div>
         </a>
-        <a href="../../#projects" class="desktop-icon">
+        <a href="/#projects" class="desktop-icon">
             <div class="icon-image">📁</div>
             <div class="icon-label">Projects</div>
         </a>
-        <a href="../../#about" class="desktop-icon">
+        <a href="/#about" class="desktop-icon">
             <div class="icon-image">📝</div>
             <div class="icon-label">Notes</div>
         </a>
-        <a href="../../#contact" class="desktop-icon">
+        <a href="/#contact" class="desktop-icon">
             <div class="icon-image">✉️</div>
             <div class="icon-label">Contact</div>
         </a>
@@ -84,6 +84,6 @@
         </div>
     </main>
 
-    <script src="../../js/main.js" defer></script>
+    <script type="module" src="/scripts/global.js"></script>
 </body>
 </html>
