--- notes/about.md
@@ -1,49 +1,0 @@
----
-title: About Me
-description: Software & Hardware Engineer's Journey
-author: Paul Junsuk Han
-date: 2024-01-03
-tags: [engineering, software, hardware, AI]
-category: profile
-lastModified: 2024-01-03
-visibility: public
----
-
-<!--
-SEO Keywords:
-software engineer, hardware engineer, full stack developer, AI engineer, machine learning engineer,
-technical architect, system design, neural networks, hardware optimization, engineering philosophy,
-Paul Han, Paul Junsuk Han, software development, hardware development, engineering principles
--->
-
-SWE, HWE, AI, builder, question asker. <h6>#40 <a href="https://guzey.com/people/" target="_blank">Alexey Guzey's People who are going to change the world</a></h6>
-
-<!--
-Core competencies: software, hardware design, distributed systems, machine learning
-scalable solutions, performance optimization, technical leadership
--->
-
-
-
-
-<div class="markdown-figure" style="width: 300px; margin: 0;">
-
-![3D printed brain](/assets/images/image.png "3D printed from my MRI AI segmentation")
-
-</div>
-
-<h5> 3D printed my brain using MRI scan. Learned a lot. </h5>
-
-<!--
-Research areas: neural network optimization, hardware acceleration, 
-AI model compression, efficient inference, edge computing, embedded AI
--->
-
-
-
-
-<h3>Big Interests</h3>
-AI Guardian - AI-Human intelligence augmentation. tFUS/BCI/neurotech. Silent speech interfaces.
-
-
-<h6>no libraries were involved in the making of this website. </h6>
