--- projects/neural-network/body.md
@@ -1,46 +1,0 @@
----
-title: Neural Network Implementation Body
-description: Technical implementation details
-author: Paul Junsuk Han
-date: 2024-01-03
-tags: [neural-network, implementation, architecture]
-category: project
-lastModified: 2024-01-03
-visibility: public
----
-
-The visualization engine uses WebGL for hardware-accelerated rendering of complex network structures. React manages the UI state and user interactions, while Three.js provides the 3D rendering capabilities.
-
-<div class="sketch-list">
-<h3>Architecture Highlights</h3>
-<ul>
-<li>Core Architecture
-    <ul>
-    <li>WebGL-based rendering pipeline</li>
-    <li>React component hierarchy</li>
-    <li>Three.js scene management</li>
-    </ul>
-</li>
-<li>Performance Optimizations
-    <ul>
-    <li>Custom WebGL shaders for node rendering</li>
-    <li>Instanced geometry for large networks</li>
-    <li>Optimized graph layout algorithms</li>
-    </ul>
-</li>
-<li>Interactive Features
-    <ul>
-    <li>Real-time data flow visualization</li>
-    <li>Node inspection and modification</li>
-    <li>Network architecture manipulation</li>
-    </ul>
-</li>
-<li>Development Tools
-    <ul>
-    <li>Modular component architecture</li>
-    <li>Hot module replacement</li>
-    <li>Comprehensive testing suite</li>
-    </ul>
-</li>
-</ul>
-</div>
