--- projects/neural-network/config.html
@@ -1,22 +1,0 @@
-<!-- Project Configuration -->
-<div class="project-config" hidden>
-    <!-- Project Metadata -->
-    <div id="project-meta">
-        <span id="project-title">Neural Network Visualizer - Project Details</span>
-        <span id="project-github">https://github.com/pauljunsukhan/neural-network</span>
-        <span id="project-date">1.02.2024</span>
-    </div>
-
-    <!-- Project Sections -->
-    <div id="project-sections">
-        <!-- Overview Section -->
-        <section id="overview" data-markdown="./overview.md">
-            <h2>Project Overview</h2>
-        </section>
-
-        <!-- Body Section -->
-        <section id="body" data-markdown="./body.md">
-            <h2>Body</h2>
-        </section>
-    </div>
-</div>
