--- projects/neural-network/overview.md
@@ -1,32 +1,0 @@
----
-title: Neural Network Implementation
-description: Technical details of the Neural Network Visualizer
-author: Paul Junsuk Han
-date: 2024-01-03
-tags: [neural-network, webgl, react, threejs]
-category: project
-lastModified: 2024-01-03
-visibility: public
----
-
-The visualization engine uses WebGL for hardware-accelerated rendering of complex network structures. React manages the UI state and user interactions, while Three.js provides the 3D rendering capabilities.
-
-<div class="sketch-list">
-<h3>Architecture Highlights</h3>
-<ul>
-<li>Custom WebGL shaders for efficient node rendering
-    <ul>
-    <li>Fragment shaders for node coloring</li>
-    <li>Vertex shaders for positioning</li>
-    </ul>
-</li>
-<li>Optimized graph layout algorithms
-    <ul>
-    <li>Force-directed layout</li>
-    <li>Hierarchical tree layout</li>
-    </ul>
-</li>
-<li>Real-time data flow visualization</li>
-<li>Modular component architecture</li>
-</ul>
-</div>
