--- projects/quantum-circuit/index.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="Quantum Circuit Simulator - Browser-based quantum computing simulation tool">
-    <title>Quantum Circuit Simulator | Paul Junsuk Han</title>
-    <link rel="stylesheet" href="../../styles/main.css">
-</head>
-<body>
-    <!-- Classic Mac Menu Bar -->
-    <div class="global-menu-bar">
-        <span class="apple-menu">🍎</span>
-        <div class="menu-divider"></div>
-        <span class="menu-item">File</span>
-        <span class="menu-item">Edit</span>
-        <span class="menu-item">View</span>
-        <span class="menu-item">Special</span>
-    </div>
-
-    <!-- Main Project Window -->
-    <main class="mac-window">
-        <div class="window-title-bar">
-            <div class="window-controls">
-                <div class="window-button"></div>
-                <div class="window-button"></div>
-                <div class="window-button"></div>
-            </div>
-            <div class="window-title">Quantum Circuit Simulator - Project Details</div>
-        </div>
-
-        <div class="content">
-            <!-- Project Header -->
-            <div class="notebook-header">
-                <div class="date">Status: <span class="typewriter">Beta Testing</span></div>
-                <div class="subject">Project: <span class="typewriter">Quantum Computing Simulation</span></div>
-            </div>
-
-            <!-- Project Overview -->
-            <section class="technical-sketch">
-                <h2>Project Overview</h2>
-                <p>A browser-based quantum circuit simulator that allows users to design, test, and visualize quantum circuits. Built with Python and WebAssembly, this tool makes quantum computing concepts accessible to students and researchers.</p>
-                
-                <div class="sketch-list">
-                    <h3>Technical Specifications</h3>
-                    <ul>
-                        <li>Backend: Python, Qiskit</li>
-                        <li>Frontend: WebAssembly, Three.js</li>
-                        <li>Computation: NumPy, SciPy</li>
-                        <li>Visualization: D3.js</li>
-                    </ul>
-                </div>
-            </section>
-
-            <!-- Key Features -->
-            <section class="technical-sketch">
-                <h2>Key Features</h2>
-                <ul class="sketch-list">
-                    <li>Visual quantum circuit designer</li>
-                    <li>Real-time state vector visualization</li>
-                    <li>Common quantum gate library</li>
-                    <li>Measurement and error simulation</li>
-                    <li>Circuit optimization suggestions</li>
-                </ul>
-            </section>
-
-            <!-- Implementation Details -->
-            <section class="technical-sketch">
-                <h2>Implementation Details</h2>
-                <p>The simulator leverages WebAssembly to run complex quantum computations directly in the browser. Python code is compiled to WebAssembly using Emscripten, while the frontend provides an intuitive interface for circuit design.</p>
-                
-                <div class="sketch-list">
-                    <h3>Architecture Highlights</h3>
-                    <ul>
-                        <li>Quantum state vector simulation</li>
-                        <li>Gate-level optimization</li>
-                        <li>Interactive circuit visualization</li>
-                        <li>Real-time computation feedback</li>
-                    </ul>
-                </div>
-            </section>
-
-            <!-- Project Links -->
-            <section class="technical-sketch">
-                <h2>Project Resources</h2>
-                <ul class="sketch-list">
-                    <li>Documentation: <a href="#" target="_blank">User Guide</a></li>
-                    <li>Tutorial: <a href="#" target="_blank">Getting Started</a></li>
-                    <li>Examples: <a href="#" target="_blank">Sample Circuits</a></li>
-                </ul>
-            </section>
-        </div>
-    </main>
-
-    <script src="../../js/main.js" defer></script>
-</body>
-</html>
