--- projects/template.html
@@ -1,117 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="[Project Name] - [Brief Description]">
-    <title>[Project Name] | Paul Junsuk Han</title>
-    <link rel="stylesheet" href="../../styles/main.css">
-</head>
-<body>
-    <!-- Classic Mac Menu Bar -->
-    <div class="global-menu-bar">
-        <span class="apple-menu">🍎</span>
-        <div class="menu-divider"></div>
-        <span class="menu-item">File</span>
-        <span class="menu-item">Edit</span>
-        <span class="menu-item">View</span>
-        <span class="menu-item">Special</span>
-    </div>
-
-    <!-- Desktop Icons -->
-    <div class="desktop-icons">
-        <a href="../../" class="desktop-icon">
-            <div class="icon-image">💾</div>
-            <div class="icon-label">Home</div>
-        </a>
-        <a href="../../#projects" class="desktop-icon">
-            <div class="icon-image">📁</div>
-            <div class="icon-label">Projects</div>
-        </a>
-        <a href="../../#about" class="desktop-icon">
-            <div class="icon-image">📝</div>
-            <div class="icon-label">Notes</div>
-        </a>
-        <a href="../../#contact" class="desktop-icon">
-            <div class="icon-image">✉️</div>
-            <div class="icon-label">Contact</div>
-        </a>
-    </div>
-
-    <!-- Main Project Window -->
-    <main class="mac-window">
-        <div class="window-title-bar">
-            <div class="window-controls">
-                <div class="window-button"></div>
-                <div class="window-button"></div>
-                <div class="window-button"></div>
-            </div>
-            <div class="window-title">[Project Name] - Project Details</div>
-            <a href="https://github.com/pauljunsukhan/[project-repo]" class="window-link" target="_blank">Github</a>
-        </div>
-
-        <div class="content">
-            <!-- Project Header -->
-            <div class="notebook-header">
-                <div class="date">Status: <span class="typewriter">[Project Status]</span></div>
-                <div class="subject">Project: <span class="typewriter">[Project Type]</span></div>
-            </div>
-
-            <!-- Project Overview -->
-            <section class="technical-sketch">
-                <h2>Project Overview</h2>
-                <p>[Project description - A detailed paragraph explaining the project's purpose, goals, and significance.]</p>
-                
-                <div class="sketch-list">
-                    <h3>Technical Specifications</h3>
-                    <ul>
-                        <li>Frontend: [Technologies]</li>
-                        <li>Backend: [Technologies]</li>
-                        <li>Core Tech: [Key Technologies]</li>
-                        <li>Build/Deploy: [Tools]</li>
-                    </ul>
-                </div>
-            </section>
-
-            <!-- Key Features -->
-            <section class="technical-sketch">
-                <h2>Key Features</h2>
-                <ul class="sketch-list">
-                    <li>[Feature 1 with brief description]</li>
-                    <li>[Feature 2 with brief description]</li>
-                    <li>[Feature 3 with brief description]</li>
-                    <li>[Feature 4 with brief description]</li>
-                </ul>
-            </section>
-
-            <!-- Implementation Details -->
-            <section class="technical-sketch">
-                <h2>Implementation Details</h2>
-                <p>[Technical overview - Explain the core technical implementation, architecture decisions, and notable technical challenges solved.]</p>
-                
-                <div class="sketch-list">
-                    <h3>Architecture Highlights</h3>
-                    <ul>
-                        <li>[Technical highlight 1]</li>
-                        <li>[Technical highlight 2]</li>
-                        <li>[Technical highlight 3]</li>
-                        <li>[Technical highlight 4]</li>
-                    </ul>
-                </div>
-            </section>
-
-            <!-- Project Links -->
-            <section class="technical-sketch">
-                <h2>Project Resources</h2>
-                <ul class="sketch-list">
-                    <li>Github Repository: <a href="https://github.com/pauljunsukhan/[project-repo]" target="_blank">View Source</a></li>
-                    <li>Documentation: <a href="[docs-url]" target="_blank">Technical Docs</a></li>
-                    <li>Live Demo: <a href="[demo-url]" target="_blank">Try it Out</a></li>
-                </ul>
-            </section>
-        </div>
-    </main>
-
-    <script src="../../js/main.js" defer></script>
-</body>
-</html>
