+++ scripts/encode_personaldata.js
