--- sitemap.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
-        xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
-    <url>
-        <loc>https://pauljunsukhan.com/</loc>
-        <lastmod>2024-01-02</lastmod>
-        <changefreq>weekly</changefreq>
-        <priority>1.0</priority>
-        <image:image>
-            <image:loc>https://pauljunsukhan.com/assets/site/social-preview.webp</image:loc>
-            <image:title>Paul Han - Portfolio</image:title>
-        </image:image>
-    </url>
-    <url>
-        <loc>https://pauljunsukhan.com/projects/neural-network/</loc>
-        <lastmod>2024-01-02</lastmod>
-        <changefreq>monthly</changefreq>
-        <priority>0.8</priority>
-    </url>
-    <url>
-        <loc>https://pauljunsukhan.com/projects/portfolio-website/</loc>
-        <lastmod>2024-01-02</lastmod>
-        <changefreq>monthly</changefreq>
-        <priority>0.8</priority>
-    </url>
-</urlset>
