+++ styles/desktop.css
@@ -1,0 +1,90 @@
+/* ===============================
+ * 2) desktop.css
+ * =============================== */
+
+/* Desktop Icons & Layout */
+.desktop-icons {
+    position: fixed;
+    top: 40px;
+    right: 20px;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    z-index: 1;
+    transition: transform 0.3s ease;
+}
+.desktop-icon {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    width: 64px;
+    cursor: pointer;
+    transition: transform 0.2s ease;
+    text-decoration: none;
+    color: inherit;
+}
+.desktop-icon:hover {
+    transform: scale(1.05);
+}
+.icon-image {
+    width: 32px;
+    height: 32px;
+    margin-bottom: 5px;
+    background: var(--menu-bg);
+    border: 1px solid var(--window-border);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 20px;
+}
+.icon-label {
+    font-family: "Courier New", monospace;
+    font-size: 11px;
+    color: var(--window-border);
+    background: var(--menu-bg);
+    padding: 2px 4px;
+    border-radius: 0;
+    border: 1px solid var(--window-border);
+}
+
+/* Mobile toggling for .desktop-icons */
+@media (max-width: 768px) {
+    .desktop-icons {
+        right: 5px;
+        gap: 10px;
+        transform: translateX(100%);
+    }
+    .desktop-icons.active {
+        transform: translateX(0);
+    }
+    .desktop-icon {
+        transform: scale(0.8);
+    }
+    .icon-label {
+        font-size: 10px;
+    }
+}
+
+/* The mobile “⋮” toggle button */
+.menu-toggle {
+    display: none;
+    position: fixed;
+    top: 50%;
+    right: 0;
+    width: 20px;
+    height: 40px;
+    background: var(--menu-bg);
+    border: 1px solid var(--window-border);
+    border-right: none;
+    border-radius: 4px 0 0 4px;
+    cursor: pointer;
+    z-index: 2;
+    align-items: center;
+    justify-content: center;
+}
+@media (max-width: 768px) {
+    .menu-toggle {
+        display: flex;
+    }
+}
