+++ styles/socials.css
@@ -1,0 +1,101 @@
+/* ===============================
+ * socials.css
+ * =============================== */
+
+/* Social Links Grid */
+.social-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
+  gap: 1rem;
+  padding: 1rem;
+  list-style: none;
+  margin: 0;
+}
+
+.social-link {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.5rem;
+  text-decoration: none;
+  color: var(--primary-color);
+  background: var(--window-bg);
+  border: 1px solid var(--window-border);
+  border-radius: 4px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.social-link:hover {
+  background: var(--menu-bg);
+  transform: translateY(-1px);
+  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+}
+
+.social-icon {
+  font-size: 1.2em;
+  min-width: 1.5em;
+  text-align: center;
+}
+
+.social-label {
+  font-family: "Courier New", monospace;
+  font-size: 0.9em;
+}
+
+/* Mac Dialog */
+.mac-dialog {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%) scale(0.9);
+  background: var(--window-bg);
+  border: 1px solid var(--window-border);
+  border-radius: 6px;
+  min-width: 300px;
+  max-width: 90vw;
+  opacity: 0;
+  visibility: hidden;
+  transition: all 0.2s ease;
+  z-index: 2000;
+}
+
+.mac-dialog.active {
+  transform: translate(-50%, -50%) scale(1);
+  opacity: 1;
+  visibility: visible;
+}
+
+.mac-dialog .content {
+  padding: 1rem;
+}
+
+.protected-content {
+  font-family: "Courier New", monospace;
+  background: rgba(0,0,0,0.05);
+  padding: 0.5rem;
+  border-radius: 4px;
+  word-break: break-all;
+}
+
+/* Responsive Adjustments */
+@media (max-width: 480px) {
+  .social-grid {
+    grid-template-columns: repeat(auto-fill, minmax(100px, 1fr));
+    gap: 0.5rem;
+    padding: 0.5rem;
+  }
+
+  .social-link {
+    padding: 0.4rem;
+  }
+
+  .social-label {
+    font-size: 0.8em;
+  }
+
+  .mac-dialog {
+    min-width: 250px;
+    margin: 0 1rem;
+  }
+}
